+++ src/components/ui/Label.h
@@ -19,8 +19,8 @@ public:
     Label(
         const char* msg,
         const Position& position,
-        int width,
-        int height,
+        float width,
+        float height,
         uint8_t fontSize = 24,
         const Color& color = Color(255, 255, 255, 255),
         const char* pathToFont = nullptr,
